Split into abstract benchmark
